engines/backend/requirements.txt
# Core framework
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.3.0
starlette==0.27.0

# Database
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.7

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Task queue and caching
celery==5.3.4
redis==4.6.0
flower==2.0.1

# API features
email-validator==2.0.0.post2
httpx==0.24.1
orjson==3.9.5
ujson==5.8.0

# Data processing
pyyaml==6.0.1
numpy==1.26.0
pandas==2.1.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
aiofiles==23.2.1
